implemented authorization

+++ service/api/deletePhoto.go
@@ -14,6 +14,13 @@ func (rt *_router) deletePhoto(w http.ResponseWriter, r *http.Request, ps httpro
 		w.WriteHeader(http.StatusBadRequest)
 		return
 	}
+
+	isAuth := rt.db.IsAuthorized(getToken(r.Header.Get("Authorization")))
+	if !isAuth {
+		w.WriteHeader(http.StatusUnauthorized)
+		return
+	}
+
 	err = rt.db.DeletePhoto(uint64(id))
 	if err != nil {
 		w.WriteHeader(http.StatusNotFound)
